ASACORE-1526 Fixes to address /W4 compile issues.
Most common warnings to be resolved were:
1. Unreferenced formal parameter. Introduced UNREFERENCED_PARAMETER macros.
2. Assignment operator could not be generated - introduced private assigment operators
3. Conditional expression is constant - loop implemented like 'while(0){}' or 'while(true){}'.
Used for(;;){} or for(;;){break;} instead.
4. Conversion from 'int' to 'unit32_t'('uint8_t', 'DWORD' etc), signed/unsigned mismatch.
Introduced casts.

+++ alljoyn_c/unit_test/BusAttachmentTest.cc
@@ -84,7 +84,7 @@ TEST(BusAttachmentTest, isstarted_isstopping) {
 
 TEST(BusAttachmentTest, getconcurrency) {
     alljoyn_busattachment bus = NULL;
-    unsigned int concurrency = -1;
+    unsigned int concurrency = (unsigned int)-1;
     bus = alljoyn_busattachment_create("BusAttachmentTest", QCC_TRUE);
 
     concurrency = alljoyn_busattachment_getconcurrency(bus);
@@ -94,7 +94,7 @@ TEST(BusAttachmentTest, getconcurrency) {
     alljoyn_busattachment_destroy(bus);
 
     bus = NULL;
-    concurrency = -1;
+    concurrency = (unsigned int)-1;
 
     bus = alljoyn_busattachment_create_concurrency("BusAttachmentTest", QCC_TRUE, 8);
